level change
I changed the level to what we originally discussed, as far as our hero traversiing the rooftops. I made it to where the player is forced to use the hook with every move. hope you like. that should give a better idea of how far we want the hook to shoot out